Introduce/enhance delete operations

+++ src/main/kotlin/com/mibe/iot/thinker/device/application/DeleteDeviceService.kt
@@ -1,9 +1,10 @@
 package com.mibe.iot.thinker.device.application
 
 import com.mibe.iot.thinker.device.application.port.from.DeleteDevicePort
+import com.mibe.iot.thinker.device.application.port.from.GetDevicePort
 import com.mibe.iot.thinker.device.application.port.to.DeleteDeviceUseCase
+import com.mibe.iot.thinker.device.application.port.to.exception.DeviceNotFoundException
 import org.springframework.stereotype.Service
-import reactor.core.publisher.Mono
 
 /**
  * Delete device service
@@ -13,10 +14,16 @@ import reactor.core.publisher.Mono
  */
 @Service
 class DeleteDeviceService(
-    private val deleteDevicePort: DeleteDevicePort
+    private val deleteDevicePort: DeleteDevicePort,
+    private val getDevicePort: GetDevicePort
 ) : DeleteDeviceUseCase {
 
-    override fun deleteDevice(id: String): Mono<Void> {
-        return deleteDevicePort.deleteDevice(id)
+    override suspend fun deleteDevice(id: String) {
+        if (getDevicePort.existsWithId(id)) {
+            deleteDevicePort.deleteDevice(id)
+        } else {
+            throw DeviceNotFoundException(id)
+        }
+
     }
 }
